move duplicate modal to saved_objects for reuse
