+++ src/plugins/dashboard/public/application/components/dashboard_top_nav/top_nav/index.ts
@@ -6,6 +6,5 @@
 export * from './get_top_nav_config';
 export * from './options';
 export * from './save_modal';
-export * from './show_duplicate_modal';
 export * from './show_options_popover';
 export * from './top_nav_ids';
