+++ src/plugins/saved_objects_management/public/index.ts
@@ -51,12 +51,13 @@ export {
   processImportResponse,
   FailedImport,
   duplicateSavedObjects,
+  getSavedObjectLabel,
   getWorkspacesWithWritePermission,
 } from './lib';
-export { SavedObjectsDuplicateModal, DuplicateMode } from './management_section';
 export { SavedObjectRelation, SavedObjectWithMetadata, SavedObjectMetadata } from './types';
 export { SAVED_OBJECT_DELETE_TRIGGER, savedObjectDeleteTrigger } from './triggers';
 export { SavedObjectDeleteContext } from './ui_actions_bootstrap';
+export { SAVED_OBJECT_TYPE_WORKSPACE } from './constants';
 
 export function plugin(initializerContext: PluginInitializerContext) {
   return new SavedObjectsManagementPlugin();
