+++ src/plugins/saved_objects_management/public/management_section/objects_table/components/index.ts
@@ -32,4 +32,3 @@ export { Header } from './header';
 export { Table } from './table';
 export { Flyout } from './flyout';
 export { Relationships } from './relationships';
-export { SavedObjectsDuplicateModal } from './duplicate_modal';
